machines: Pull volume usage into a helper function
Pull code which finds out if storage volumes are used by any virtual
machine from storagePoolVolumesTab component to helpers.
